fix: calculating pid temperature in float, fixed #23

+++ src/RegulatorTask.h
@@ -3,7 +3,7 @@
 #include <PIDtuner.h>
 
 Equitherm etRegulator;
-GyverPID<int, unsigned long> pidRegulator(0, 0, 0);
+GyverPID<float, unsigned long> pidRegulator(0, 0, 0);
 PIDtuner pidTuner;
